Move resize trigger out to application root
