+++ package.json
@@ -22,7 +22,8 @@
     "blick": "^1.1.0",
     "domready": "^1.0.8",
     "global": "^4.3.0",
-    "gutentag": "^2.0.0"
+    "gutentag": "^2.0.0",
+    "ndim": "^1.0.0"
   },
   "translators": {
     "html": "gutentag/translate-html"
